@@ -30,13 +30,9 @@
 #include "gfx/hud.h"
 #include "vs_globals.h"
 
-#if defined(__APPLE__) && defined(__MACH__)
-    #include <GLUT/glut.h>
-    #include <OpenGL/glext.h>
-#else
-    #include <GL/glut.h>
-    #include <GL/glext.h>
-#endif
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
+#include <glut.h>
+#include <glext.h>
 
 #include <iostream>
 //#include <stdlib>
