@@ -93,8 +93,8 @@ static float EstimateGain(const Vector &pos, const float gain) {
     float listener_size = sqrt(mylistener.rsize);
     float distance = (AUDListenerLocation() - pos.Cast()).Magnitude()
             - listener_size
-            - game_options()->audio_ref_distance;
-    float ref = game_options()->audio_ref_distance;
+            - vs_options::instance().audio_ref_distance;
+    float ref = vs_options::instance().audio_ref_distance;
     float rolloff = 1.0f;
     final_gain *= (distance <= 0) ? 1.f : float(ref / (ref + rolloff * distance));
 
@@ -115,7 +115,7 @@ char AUDQueryAudability(const int sound, const Vector &pos, const Vector &vel, c
         mag = 0;
         return 1;
     }
-    if (!(mag < game_options()->audio_max_distance * game_options()->audio_max_distance)) {
+    if (!(mag < vs_options::instance().audio_max_distance * vs_options::instance().audio_max_distance)) {
         return 0;
     }
     unsigned int hashed = hash_sound(sounds[sound].buffer);
