@@ -539,7 +539,7 @@ int AUDCreateSoundWAV(const std::string &s, const bool music, const bool LOOP) {
 #ifdef SOUND_DEBUG
     VS_LOG(trace, "AUDCreateSoundWAV:: ");
 #endif
-    if ((game_options()->Music && !music) || (game_options()->Music && music)) {
+    if ((vs_options::instance().Music && !music) || (vs_options::instance().Music && music)) {
         ALuint *wavbuf = NULL;
         std::string hashname;
         if (!music) {
@@ -590,7 +590,7 @@ int AUDCreateMusicWAV(const std::string &s, const bool LOOP) {
 int AUDCreateSoundMP3(const std::string &s, const bool music, const bool LOOP) {
 #ifdef HAVE_AL
     assert(0);
-    if ((game_options()->Music && !music) || (game_options()->Music && music)) {
+    if ((vs_options::instance().Music && !music) || (vs_options::instance().Music && music)) {
         VSFile f;
         VSError error = f.OpenReadOnly(s.c_str(), SoundFile);
         bool shared = (error == Shared);
@@ -738,8 +738,8 @@ void AUDAdjustSound(const int sound, const QVector &pos, const Vector &vel) {
             alSourcei(sounds[sound].source, AL_SOURCE_RELATIVE, relative);
             if (!relative) {
                 // Set rolloff factrs
-                alSourcef(sounds[sound].source, AL_MAX_DISTANCE, scalepos * game_options()->audio_max_distance);
-                alSourcef(sounds[sound].source, AL_REFERENCE_DISTANCE, scalepos * game_options()->audio_ref_distance);
+                alSourcef(sounds[sound].source, AL_MAX_DISTANCE, scalepos * vs_options::instance().audio_max_distance);
+                alSourcef(sounds[sound].source, AL_REFERENCE_DISTANCE, scalepos * vs_options::instance().audio_ref_distance);
                 alSourcef(sounds[sound].source, AL_ROLLOFF_FACTOR, 1.f);
             }
         }
