@@ -25,11 +25,8 @@
 #define VEGA_STRIKE_ENGINE_CMD_AI_FLY_KEYBOARD_H
 
 #include "in.h"
-#if defined(__APPLE__) && defined(__MACH__)
-#include <GLUT/glut.h>
-#else
-#include <GL/glut.h>
-#endif
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
+#include <glut.h>
 #include "flybywire.h"
 
 class FlyByKeyboard : public FlyByWire {
