@@ -1,10 +1,8 @@
-/**
+/*
  * mesh_server.cpp
  *
- * Copyright (c) 2001-2002 Daniel Horn
- * Copyright (c) 2002-2019 pyramid3d and other Vega Strike Contributors
- * Copyright (c) 2019-2021 Stephen G. Tuggy, and other Vega Strike Contributors
- * Copyright (C) 2022 Stephen G. Tuggy
+ * Copyright (c) 2001-2024 Daniel Horn, pyramid3d, Stephen G. Tuggy,
+ * Benjamen R. Meyer, Roy Falk, and other Vega Strike Contributors
  *
  * https://github.com/vegastrike/Vega-Strike-Engine-Source
  *
@@ -17,7 +15,7 @@
  *
  * Vega Strike is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
@@ -120,17 +118,10 @@ Mesh::~Mesh() {
             meshHashTable.Delete(hash_name);
         }
         vector<Mesh *> *hashers = bfxmHashTable.Get(hash_name);
-        vector<Mesh *>::iterator finder;
-        if (hashers) {
-            for (int i = hashers->size() - 1; i >= 0; --i) {
-                if ((*hashers)[i] == this) {
-                    hashers->erase(hashers->begin() + i);
-                    if (hashers->empty()) {
-                        bfxmHashTable.Delete(hash_name);
-                        delete hashers;
-                    }
-                }
-            }
+        bool hashers_was_deleted = false;
+        remove_all_references_to(this, hashers, true, hashers_was_deleted);
+        if (hashers_was_deleted) {
+            VS_LOG_AND_FLUSH(debug, "hashers was deleted");
         }
         orig->refcount--;
         //printf ("orig refcount: %d",refcount);
