@@ -36,10 +36,17 @@
 #include "universe.h"
 #include "vs_logging.h"
 
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
 #if defined(__APPLE__) && defined(__MACH__)
-#include <OpenGL/gl.h>
+#   include <gl.h>
+#elif defined (_WIN32) || defined (__CYGWIN__)
+#   ifndef NOMINMAX
+#       define NOMINMAX
+#   endif //tells VCC not to generate min/max macros
+#   include <windows.h>
+#   include <gl.h>
 #else
-#include <GL/gl.h>
+#   include <gl.h>
 #endif
 
 #define SINX 1
