@@ -25,13 +25,9 @@
 
 #include "vsimage.h"
 
-#if defined(__APPLE__) && defined(__MACH__)
-#   include <GLUT/glut.h>
-#   include <OpenGL/glext.h>
-#else
-#   include <GL/glut.h>
-#   include <GL/glext.h>
-#endif
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
+#include <glut.h>
+#include <glext.h>
 
 #include "cmd/unit_generic.h"
 
