@@ -93,17 +93,20 @@ struct GFXStats
 
 #ifndef _WIN32
 //#define GL_GLEXT_PROTOTYPES
-
 #endif
+
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
 #if defined (_WIN32) || defined (__CYGWIN__)
 #ifndef NOMINMAX
 #define NOMINMAX
 #endif //tells VCC not to generate min/max macros
 #include <windows.h>
-#include <GL/gl.h>
+#include <gl.h>
 #endif
+
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
 #if defined(__APPLE__) && defined(__MACH__)
-#include <GLUT/glut.h>
+#include <glut.h>
 //#if defined( GL_INIT_CPP) || defined( GL_MISC_CPP) || defined( GL_STATE_CPP)
 #if defined (GL_ARB_vertex_program) && defined (GL_ARB_fragment_program)
 #define OSX_AT_LEAST_10_4
@@ -112,15 +115,16 @@ struct GFXStats
 #endif
 #define GL_GLEXT_PROTOTYPES
 //#endif
-#include <OpenGL/glext.h>
+#include <glext.h>
 #else
 #define __glext_h_
-#include <GL/glut.h>
+#include <glut.h>
 #include "gl_undefined_extensions.h"
 #undef __glext_h_
 
-#include <GL/glext.h>
+#include <glext.h>
 #endif
+
 #ifdef _WIN32
 #   define GL_TEXTURE0_ARB 0x84C0
 #   define GL_TEXTURE1_ARB 0x84C1
