@@ -54,13 +54,14 @@
 #include <windows.h>
 #endif
 #define GL_GLEXT_PROTOTYPES 1
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
 #if defined(__APPLE__) && defined(__MACH__)
-    #include <OpenGL/gl.h>
-    #include <OpenGL/glext.h>
-    #include <dlfcn.h>
+#   include <gl.h>
+#   include <glext.h>
+#   include <dlfcn.h>
 #else
-    #include <GL/gl.h>
-    #include <GL/glext.h>
+#   include <gl.h>
+#   include <glext.h>
 #endif
 #ifdef GL_EXT_compiled_vertex_array
 # ifndef PFNGLLOCKARRAYSEXTPROC
@@ -75,7 +76,7 @@
     # define GLX_GLXEXT_PROTOTYPES 1
     # define GLX_GLXEXT_LEGACY 1
     # include <GL/glx.h>
-    # include <GL/glext.h>
+    # include <glext.h>
 #endif
 #endif
 
@@ -149,7 +150,7 @@ typedef void ( *(*get_gl_proc_fptr_t)(const GLubyte *))();
     #if defined(__HAIKU__)
         typedef char * GET_GL_PTR_TYP;
         #define GET_GL_PROC glutGetProcAddress
-    #elif defined (__MACOSX__)
+    #elif defined (__APPLE__) && defined (__MACH__)
         typedef const char * GET_GL_PTR_TYP;
         void * vega_dlsym(GET_GL_PTR_TYP function_name) {
             return dlsym(RTLD_SELF, function_name);
@@ -168,8 +169,9 @@ CG_Cloak *cloak_cg = new CG_Cloak();
 /* CENTRY */
 int vsExtensionSupported(const char *extension) {
     static const GLubyte *extensions = nullptr;
-    const GLubyte *start;
-    GLubyte *where, *terminator;
+    const GLubyte *start = nullptr;
+    GLubyte *where = nullptr;
+    GLubyte *terminator = nullptr;
 
     /* Extension names should not have spaces. */
     where = (GLubyte *) strchr(extension, ' ');
@@ -178,6 +180,10 @@ int vsExtensionSupported(const char *extension) {
     }
     if (!extensions) {
         extensions = glGetString(GL_EXTENSIONS);
+        if (!extensions) {
+            VS_LOG_AND_FLUSH(serious_warning, "glGetString(GL_EXTENSIONS) returned NULL!");
+            return false;
+        }
     }
     /* It takes a bit of care to be fool-proof about parsing the
      *  OpenGL extensions string.  Don't be fooled by sub-strings,
@@ -245,11 +251,15 @@ bool vsVendorMatch(const char *vendor) {
 void init_opengl_extensions() {
     const unsigned char *extensions = glGetString(GL_EXTENSIONS);
 
-    VS_LOG(trace, (boost::format("OpenGL Extensions supported: %1%") % extensions));
+    if (extensions) {
+        VS_LOG(trace, (boost::format("OpenGL Extensions supported: %1%") % extensions));
+    } else {
+        VS_LOG(serious_warning, "OpenGL Extensions supported: none!");
+    }
 
 #ifndef NO_COMPILEDVERTEXARRAY_SUPPORT
     if (vsExtensionSupported("GL_EXT_compiled_vertex_array")
-            && game_options()->LockVertexArrays) {
+            && vs_options::instance().LockVertexArrays) {
 #if defined(__APPLE__) && defined(__MACH__)
 #ifndef __APPLE_PANTHER_GCC33_CLI__
 #if defined (glLockArraysEXT) && defined (glUnlockArraysEXT)
@@ -431,7 +441,7 @@ void init_opengl_extensions() {
 #ifdef GL_FOG_DISTANCE_MODE_NV
     if (vsExtensionSupported("GL_NV_fog_distance")) {
         VS_LOG(trace, "OpenGL::Accurate Fog Distance supported");
-        switch (game_options()->fogdetail) {
+        switch (vs_options::instance().fogdetail) {
             case 0:
                 glFogi(GL_FOG_DISTANCE_MODE_NV, GL_EYE_PLANE_ABSOLUTE_NV);
                 break;
@@ -548,18 +558,18 @@ void init_opengl_extensions() {
 }
 
 static void initfov() {
-    g_game.fov = game_options()->fov;
-    g_game.aspect = game_options()->aspect;
-    g_game.znear = game_options()->znear;
-    g_game.zfar = game_options()->zfar;
-    g_game.detaillevel = game_options()->ModelDetail;
-    g_game.use_textures = game_options()->UseTextures;
-    g_game.use_ship_textures = game_options()->UseShipTextures;
-    g_game.use_planet_textures = game_options()->UsePlanetTextures;
-    g_game.use_logos = game_options()->UseLogos;
-    g_game.use_sprites = game_options()->UseVSSprites;
-    g_game.use_animations = game_options()->UseAnimations;
-    g_game.use_videos = game_options()->UseVideos;
+    g_game.fov = vs_options::instance().fov;
+    g_game.aspect = vs_options::instance().aspect;
+    g_game.znear = vs_options::instance().znear;
+    g_game.zfar = vs_options::instance().zfar;
+    g_game.detaillevel = vs_options::instance().ModelDetail;
+    g_game.use_textures = vs_options::instance().UseTextures;
+    g_game.use_ship_textures = vs_options::instance().UseShipTextures;
+    g_game.use_planet_textures = vs_options::instance().UsePlanetTextures;
+    g_game.use_logos = vs_options::instance().UseLogos;
+    g_game.use_sprites = vs_options::instance().UseVSSprites;
+    g_game.use_animations = vs_options::instance().UseAnimations;
+    g_game.use_videos = vs_options::instance().UseVideos;
 
     /*
      *  FILE * fp = fopen ("glsetup.txt","r");
@@ -586,17 +596,16 @@ void GFXInit(int argc, char **argv) {
     char vsicon[9] = "vega.ico";
     winsys_init(&argc, argv, &vsname[0], &vsicon[0]);
 
-
     glViewport(0, 0, g_game.x_resolution, g_game.y_resolution);
     static GFXColor clearcol = vs_config->getColor("space_background");;
-    gl_options.wireframe = game_options()->use_wireframe;
-    gl_options.max_texture_dimension = game_options()->max_texture_dimension;
-    gl_options.max_movie_dimension = game_options()->max_movie_dimension;
+    gl_options.wireframe = vs_options::instance().use_wireframe;
+    gl_options.max_texture_dimension = vs_options::instance().max_texture_dimension;
+    gl_options.max_movie_dimension = vs_options::instance().max_movie_dimension;
     bool textsupported =
             (vsExtensionSupported("GL_ARB_texture_non_power_of_two") || vsExtensionSupported("GL_ARB_texture_rectangle")
                     || vsExtensionSupported("GL_NV_texture_rectangle")) ? "true" : "false";
 
-    gl_options.rect_textures = game_options()->rect_textures ? true : textsupported;
+    gl_options.rect_textures = vs_options::instance().rect_textures ? true : textsupported;
 
     if (gl_options.rect_textures) {
         VS_LOG(trace, "RECT textures supported");
@@ -614,7 +623,7 @@ void GFXInit(int argc, char **argv) {
     bool vidsupported = (gl_options.rect_textures
                          || (vsExtensionSupported("GL_ARB_texture_non_power_of_two") && vsVendorMatch("nvidia")));
 
-    gl_options.pot_video_textures = game_options()->pot_video_textures ? true : vidsupported;
+    gl_options.pot_video_textures = vs_options::instance().pot_video_textures ? true : vidsupported;
 
     if (!gl_options.pot_video_textures && gl_options.rect_textures) {
         // Enforce max rect texture for movies, which use them
@@ -629,17 +638,17 @@ void GFXInit(int argc, char **argv) {
         VS_LOG(trace, "Using NPOT video textures");
     }*/
     // Removing gl_options soon
-    gl_options.smooth_shade = game_options()->SmoothShade;
-    gl_options.mipmap = game_options()->mipmapdetail;
-    gl_options.compression = game_options()->texture_compression;
-    gl_options.Multitexture = game_options()->reflection;
-    gl_options.smooth_lines = game_options()->smooth_lines;
-    gl_options.smooth_points = game_options()->smooth_points;
+    gl_options.smooth_shade = vs_options::instance().SmoothShade;
+    gl_options.mipmap = vs_options::instance().mipmapdetail;
+    gl_options.compression = vs_options::instance().texture_compression;
+    gl_options.Multitexture = vs_options::instance().reflection;
+    gl_options.smooth_lines = vs_options::instance().smooth_lines;
+    gl_options.smooth_points = vs_options::instance().smooth_points;
 
-    gl_options.display_lists = game_options()->displaylists;
-    gl_options.s3tc = game_options()->s3tc;
-    gl_options.ext_clamp_to_edge = game_options()->ext_clamp_to_edge;
-    gl_options.ext_clamp_to_border = game_options()->ext_clamp_to_border;
+    gl_options.display_lists = vs_options::instance().displaylists;
+    gl_options.s3tc = vs_options::instance().s3tc;
+    gl_options.ext_clamp_to_edge = vs_options::instance().ext_clamp_to_edge;
+    gl_options.ext_clamp_to_border = vs_options::instance().ext_clamp_to_border;
 
     glClearColor(clearcol.r, clearcol.g, clearcol.b, clearcol.a);
     winsys_set_reshape_func(Reshape);
@@ -731,7 +740,7 @@ void GFXInit(int argc, char **argv) {
     GFXCreateLightContext(con);
     //glutSetCursor(GLUT_CURSOR_NONE);
     /* Avoid scrambled screen on startup - Twice, for triple buffering */
-    if (game_options()->ClearOnStartup) {
+    if (vs_options::instance().ClearOnStartup) {
         glClear(GL_COLOR_BUFFER_BIT);
         winsys_swap_buffers();
         glClear(GL_COLOR_BUFFER_BIT);
