@@ -30,9 +30,10 @@
  * The only external functions it requires that aren't provided by system libs are in glut_support.h
  */
 
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
 #if defined(__APPLE__) && defined(__MACH__)
-    #include <OpenGL/gl.h>
-    #include <GLUT/glut.h>
+    #include <gl.h>
+    #include <glut.h>
 #else
 #ifdef _WIN32
     #ifndef NOMINMAX
@@ -40,9 +41,9 @@
     #endif //tells VCC not to generate min/max macros
     #include <windows.h>
 #else
-    #include <GL/gl.h>
+    #include <gl.h>
 #endif
-    #include <GL/glut.h>
+    #include <glut.h>
 #endif
 
 #include "glut_support.h"
