@@ -37,9 +37,10 @@
 //There are places where a float is converted to an int. This define takes extra steps to convert without a warning (float -> char -> int)
 #define NO_WARNINGS
 
+// See https://github.com/vegastrike/Vega-Strike-Engine-Source/pull/851#discussion_r1589254766
 #if defined(__APPLE__) && defined(__MACH__)
-    #include <OpenGL/gl.h>
-    #include <GLUT/glut.h>
+    #include <gl.h>
+    #include <glut.h>
 #else
 #ifdef _WIN32
     #ifndef NOMINMAX
@@ -47,8 +48,8 @@
     #endif //tells VCC not to generate min/max macros
     #include <windows.h>
 #endif
-    #include <GL/gl.h>
-    #include <GL/glut.h>
+    #include <gl.h>
+    #include <glut.h>
 #endif
 
 #include "glut_support.h"
