@@ -1,5 +1,7 @@
 /*
- * Copyright (C) 2001-2022 Daniel Horn, Alexander Rawass, pyramid3d,
+ * in_joystick.cpp
+ * 
+ * Copyright (C) 2001-2024 Daniel Horn, Alexander Rawass, pyramid3d,
  * Stephen G. Tuggy, and other Vega Strike contributors.
  *
  * https://github.com/vegastrike/Vega-Strike-Engine-Source
@@ -72,11 +74,11 @@ void modifyDeadZone(JoyStick *j) {
 }
 
 void modifyExponent(JoyStick *j) {
-    if ((game_options()->joystick_exponent != 1.0) && (game_options()->joystick_exponent > 0)) {
+    if ((vs_options::instance().joystick_exponent != 1.0) && (vs_options::instance().joystick_exponent > 0)) {
         for (int a = 0; a < j->nr_of_axes; a++) {
             j->joy_axis[a] =
-                    ((j->joy_axis[a] < 0) ? -pow(-j->joy_axis[a], game_options()->joystick_exponent) : pow(j->joy_axis[a],
-                            game_options()->
+                    ((j->joy_axis[a] < 0) ? -pow(-j->joy_axis[a], vs_options::instance().joystick_exponent) : pow(j->joy_axis[a],
+                            vs_options::instance().
                                     joystick_exponent));
         }
     }
@@ -146,7 +148,7 @@ JoyStick::JoyStick() {
 }
 
 int JoystickPollingRate() {
-    return game_options()->polling_rate;
+    return vs_options::instance().polling_rate;
 }
 
 void InitJoystick() {
@@ -175,7 +177,7 @@ void InitJoystick() {
     VS_LOG(info, "The names of the joysticks are:\n");
 #else
     //use glut
-    if (glutDeviceGet( GLUT_HAS_JOYSTICK ) || game_options()->force_use_of_joystick) {
+    if (glutDeviceGet( GLUT_HAS_JOYSTICK ) || vs_options::instance().force_use_of_joystick) {
         VS_LOG(info, "setting joystick functionality:: joystick online");
         glutJoystickFunc( myGlutJoystickCallback, JoystickPollingRate() );
         num_joysticks = 1;
@@ -213,11 +215,11 @@ JoyStick::JoyStick(int which) : mouse(which == MOUSE_JOYSTICK) {
     joy_buttons = 0;
 
     player = which;     //by default bind players to whichever joystick it is
-    debug_digital_hatswitch = game_options()->debug_digital_hatswitch;
+    debug_digital_hatswitch = vs_options::instance().debug_digital_hatswitch;
     if (which != MOUSE_JOYSTICK) {
-        deadzone = game_options()->deadband;
+        deadzone = vs_options::instance().deadband;
     } else {
-        deadzone = game_options()->mouse_deadband;
+        deadzone = vs_options::instance().mouse_deadband;
     };
     joy_available = 0;
     joy_x = joy_y = joy_z = 0;
@@ -298,7 +300,7 @@ void JoyStick::GetMouse(float &x, float &y, float &z, int &buttons) {
     int _mx, _my;
     GetMouseXY(_mx, _my);
     GetMouseDelta(_dx, _dy);
-    if (!game_options()->warp_mouse) {
+    if (!vs_options::instance().warp_mouse) {
         fdx = (float) (_dx = _mx - g_game.x_resolution / 2.0f);
         def_mouse_sens = 25;
         fdy = (float) (_dy = _my - g_game.y_resolution / 2.0f);
@@ -306,7 +308,7 @@ void JoyStick::GetMouse(float &x, float &y, float &z, int &buttons) {
         static std::list<mouseData> md;
         std::list<mouseData>::iterator i = md.begin();
         float ttime = getNewTime();
-        float lasttime = ttime - game_options()->mouse_blur;
+        float lasttime = ttime - vs_options::instance().mouse_blur;
         int avg = (_dx || _dy) ? 1 : 0;
         float valx = _dx;
         float valy = _dy;
@@ -340,16 +342,16 @@ void JoyStick::GetMouse(float &x, float &y, float &z, int &buttons) {
             _dx = float_to_int(valx / avg);
             _dy = float_to_int(valy / avg);
         }
-        fdx = float(valx) / game_options()->mouse_blur;
-        fdy = float(valy) / game_options()->mouse_blur;
+        fdx = float(valx) / vs_options::instance().mouse_blur;
+        fdy = float(valy) / vs_options::instance().mouse_blur;
     }
-    joy_axis[0] = fdx / (g_game.x_resolution * def_mouse_sens / game_options()->mouse_sensitivity);
-    joy_axis[1] = fdy / (g_game.y_resolution * def_mouse_sens / game_options()->mouse_sensitivity);
-    if (!game_options()->warp_mouse) {
+    joy_axis[0] = fdx / (g_game.x_resolution * def_mouse_sens / vs_options::instance().mouse_sensitivity);
+    joy_axis[1] = fdy / (g_game.y_resolution * def_mouse_sens / vs_options::instance().mouse_sensitivity);
+    if (!vs_options::instance().warp_mouse) {
         modifyDeadZone(this);
     }
-    joy_axis[0] *= game_options()->mouse_exponent;
-    joy_axis[1] *= game_options()->mouse_exponent;
+    joy_axis[0] *= vs_options::instance().mouse_exponent;
+    joy_axis[1] *= vs_options::instance().mouse_exponent;
     x = joy_axis[0];
     y = joy_axis[1];
 
