@@ -1,5 +1,7 @@
 /*
- * Copyright (C) 2001-2022 Daniel Horn, pyramid3d, Stephen G. Tuggy,
+ * lin_time.cpp
+ * 
+ * Copyright (C) 2001-2024 Daniel Horn, pyramid3d, Stephen G. Tuggy,
  * and other Vega Strike contributors.
  *
  * https://github.com/vegastrike/Vega-Strike-Engine-Source
@@ -29,15 +31,16 @@
 static double firsttime;
 VSRandom vsrandom(time(NULL));
 
-#ifdef WIN32
+#ifdef _WIN32
 #ifndef NOMINMAX
 #define NOMINMAX
 #endif //tells VCC not to generate min/max macros
 #include <windows.h>
-static LONGLONG ttime;
-static LONGLONG newtime = 0;
-static LONGLONG freq;
-static double   dblnewtime;
+static LARGE_INTEGER ttime{};
+static LARGE_INTEGER newtime{};
+static LARGE_INTEGER freq{};
+static double dblnewtime;
+static double lasttime;
 #else
 #if defined (HAVE_SDL)
 #   include <SDL2/SDL.h>
@@ -54,7 +57,7 @@ static double timecompression = 1;
 
 double getNewTime() {
 #ifdef _WIN32
-    return dblnewtime-firsttime;
+    return dblnewtime - firsttime;
 #else
     return newtime - firsttime;
 #endif
@@ -129,12 +132,6 @@ bool toggle_pause() {
 
 #ifdef _WIN32
 
-#ifndef NOMINMAX
-#define NOMINMAX
-#endif //tells VCC not to generate min/max macros
-
-#include <windows.h>
-
 void micro_sleep( unsigned int n )
 {
     Sleep( n/1000 );
@@ -144,7 +141,13 @@ void micro_sleep( unsigned int n )
 
 void micro_sleep( unsigned int n )
 {
-    (void) usleep( (useconds_t) n );
+    usleep(static_cast<useconds_t>(n));
+}
+
+#elif defined (__APPLE__) && defined (__MACH__)
+
+void micro_sleep(unsigned int n) {
+    usleep(static_cast<useconds_t>(n));
 }
 
 #else
@@ -162,9 +165,19 @@ void micro_sleep(unsigned int n) {
 #endif
 
 void InitTime() {
-#ifdef WIN32
-    QueryPerformanceFrequency( (LARGE_INTEGER*) &freq );
-    QueryPerformanceCounter( (LARGE_INTEGER*) &ttime );
+    VS_LOG(trace, "InitTime() called");
+#ifdef _WIN32
+    QueryPerformanceFrequency(&freq);
+    if (freq.QuadPart == 0) {
+        VS_LOG(serious_warning, "InitTime(): freq is zero!");
+    }
+    QueryPerformanceCounter(&ttime);
+    if (freq.QuadPart == 0) {
+        dblnewtime = static_cast<double>(ttime.QuadPart);
+    } else {
+        dblnewtime = static_cast<double>(ttime.QuadPart) / static_cast<double>(freq.QuadPart);
+    }
+    lasttime = dblnewtime - .0001;
 
 #elif defined (_POSIX_MONOTONIC_CLOCK)
     struct timespec ts;
@@ -194,10 +207,16 @@ double GetElapsedTime() {
 }
 
 double queryTime() {
-#ifdef WIN32
-    LONGLONG tmpnewtime;
-    QueryPerformanceCounter( (LARGE_INTEGER*) &tmpnewtime );
-    return ( (double) tmpnewtime )/(double) freq-firsttime;
+#ifdef _WIN32
+    LARGE_INTEGER ticks;
+    QueryPerformanceCounter(&ticks);
+    double tmpnewtime = 0;
+    if (freq.QuadPart > 0) {
+        tmpnewtime = static_cast<double>(ticks.QuadPart) / static_cast<double>(freq.QuadPart);
+    } else {
+        tmpnewtime = static_cast<double>(ticks.QuadPart);
+    }
+    return tmpnewtime - firsttime;
 #elif defined (_POSIX_MONOTONIC_CLOCK)
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
@@ -218,10 +237,18 @@ double queryTime() {
 }
 
 double realTime() {
-#ifdef WIN32
-    LONGLONG tmpnewtime;
-    QueryPerformanceCounter( (LARGE_INTEGER*) &tmpnewtime );
-    return ( (double) tmpnewtime )/(double) freq;
+#ifdef _WIN32
+    LARGE_INTEGER ticks;
+    QueryPerformanceCounter(&ticks);
+    double tmpnewtime = 0;
+    if (freq.QuadPart > 0) {
+        tmpnewtime = static_cast<double>(ticks.QuadPart) / static_cast<double>(freq.QuadPart);
+    } else {
+        tmpnewtime = static_cast<double>(ticks.QuadPart);
+    }
+    if (tmpnewtime == INFINITY) {
+        tmpnewtime = 0;
+    }
 #elif defined (_POSIX_MONOTONIC_CLOCK)
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
@@ -243,16 +270,27 @@ double realTime() {
 
 void UpdateTime() {
     static bool first = true;
-#ifdef WIN32
-    QueryPerformanceCounter( (LARGE_INTEGER*) &newtime );
-    elapsedtime = ( (double) (newtime-ttime) )/freq;
+#ifdef _WIN32
+    LARGE_INTEGER ticks;
+    QueryPerformanceCounter(&ticks);
+    lasttime = dblnewtime;
+    if (freq.QuadPart > 0) {
+        dblnewtime = static_cast<double>(ticks.QuadPart) / static_cast<double>(freq.QuadPart);
+    } else {
+        dblnewtime = static_cast<double>(ticks.QuadPart);
+    }
+    double dblttime = 0;
+    if (freq.QuadPart > 0) {
+        dblttime = static_cast<double>(ttime.QuadPart) / static_cast<double>(freq.QuadPart);
+    } else {
+        dblttime = static_cast<double>(ttime.QuadPart);
+    }
+    elapsedtime = (dblnewtime - lasttime);
     ttime = newtime;
-    if (freq == 0)
-        dblnewtime = 0.;
-    else
-        dblnewtime = ( (double) newtime )/( (double) freq );
     if (first)
+    {
         firsttime = dblnewtime;
+    }
 #elif defined(_POSIX_MONOTONIC_CLOCK)
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
@@ -281,7 +319,7 @@ void UpdateTime() {
 # error "We have no way to determine the time on this system."
 #endif
     elapsedtime *= timecompression;
-    // VS_LOG(trace, (boost::format("lin_time.cpp: UpdateTime():                                  elapsedtime after  time compression is %1%") % elapsedtime));
+    //VS_LOG(trace, (boost::format("lin_time.cpp: UpdateTime():                                  elapsedtime after  time compression is %1%") % elapsedtime));
     first = false;
 }
 
