@@ -515,7 +515,7 @@ void SaveGame::ReadNewsData(char *&buf, bool just_skip) {
 }
 
 void SaveGame::AddUnitToSave(const char *filename, int type, const char *faction, long address) {
-    if (game_options()->Drone.compare(filename)) {
+    if (vs_options::instance().Drone.compare(filename)) {
         RemoveUnitFromSave(address);
     }
 }
