@@ -1,5 +1,7 @@
 /*
- * Copyright (C) 2001-2022 Daniel Horn, pyramid3d, Stephen G. Tuggy,
+ * star_system_xml.cpp
+ * 
+ * Copyright (C) 2001-2024 Daniel Horn, pyramid3d, Stephen G. Tuggy,
  * and other Vega Strike contributors.
  *
  * https://github.com/vegastrike/Vega-Strike-Engine-Source
@@ -238,22 +240,22 @@ void parse_dual_alpha(const char *alpha, BLENDFUNC &blendSrc, BLENDFUNC &blendDs
 }
 
 int GetNumNearStarsScale() {
-    return game_options()->num_near_stars;
+    return vs_options::instance().num_near_stars;
 }
 
 int GetNumStarsScale() {
-    return game_options()->num_far_stars;
+    return vs_options::instance().num_far_stars;
 }
 
 float GetStarSpreadScale() {
-    return game_options()->star_spreading * game_options()->game_speed;
+    return vs_options::instance().star_spreading * vs_options::instance().game_speed;
 }
 
 float ScaleOrbitDist(bool autogenerated) {
     if (autogenerated) {
-        return game_options()->autogen_compactness * game_options()->runtime_compactness;
+        return vs_options::instance().autogen_compactness * vs_options::instance().runtime_compactness;
     }
-    return game_options()->runtime_compactness;
+    return vs_options::instance().runtime_compactness;
 }
 
 extern float ScaleJumpRadius(float radius);
@@ -267,14 +269,14 @@ void StarSystem::LoadXML(const string filename, const Vector &centroid, const fl
         file = filename;
     }
 
-    if (game_options()->game_speed_affects_autogen_systems) {
+    if (vs_options::instance().game_speed_affects_autogen_systems) {
         autogenerated = false;
     }
 
     xml = new Star_XML;
     xml->scale = 1;
     xml->fade = autogenerated;
-    xml->scale *= game_options()->star_system_scale;
+    xml->scale *= vs_options::instance().star_system_scale;
     xml->parentterrain = NULL;
     xml->ct = NULL;
     xml->systemcentroid = centroid;
@@ -285,7 +287,7 @@ void StarSystem::LoadXML(const string filename, const Vector &centroid, const fl
     xml->backgroundname = string("cube");
     xml->backgroundColor = GFXColor(1.0, 1.0, 1.0, 1.0);
     xml->backgroundDegamma = false;
-    xml->reflectivity = game_options()->reflectivity;
+    xml->reflectivity = vs_options::instance().reflectivity;
     xml->unitlevel = 0;
 
     VSFile other_file;
