@@ -83,7 +83,7 @@ using std::string;
 int VSFS_DEBUG() {
 //    return 3;
     if (vs_config) {
-        return (game_options()->debug_fs);
+        return (vs_options::instance().debug_fs);
     }
     return 0;
 }
@@ -346,47 +346,74 @@ std::string GetCorrectStarSysPath(const std::string &name, bool &autogenerated) 
  ***********************************************************************************************
  */
 
-FILE *vs_open(const char *filename, const char *mode) {
+FILE* vs_open(const char* filename, const char* mode) {
     if (VSFS_DEBUG() > 1) {
         cerr << "-= VS_OPEN in mode " << mode << " =- ";
     }
-    FILE *fp;
-    string fullpath = homedir + "/" + string(filename);
-    if (!use_volumes && (string(mode) == "rb" || string(mode) == "r")) {
-        string output;
-        fp = fopen(fullpath.c_str(), mode);
-        if (!fp) {
-            fullpath = string(filename);
-            output += fullpath + "... NOT FOUND\n\tTry loading : " + fullpath;
-            fp = fopen(fullpath.c_str(), mode);
+    FILE* fp = nullptr;
+    bool found = false;
+    boost::filesystem::path homedir_path{ homedir };
+    boost::filesystem::path datadir_path{ datadir };
+    boost::filesystem::path filename_path{ filename };
+    boost::filesystem::path fullpath;
+
+    if (!use_volumes && (std::string(mode) == "rb" || std::string(mode) == "r")) {
+        try {
+            fullpath = boost::filesystem::canonical(filename_path, homedir_path);
+            found = true;
+        } catch (boost::filesystem::filesystem_error& e) {
+            found = false;
         }
-        if (!fp) {
-            fullpath = datadir + "/" + string(filename);
-            output += "... NOT FOUND\n\tTry loading : " + fullpath;
-            fp = fopen(fullpath.c_str(), mode);
-        }
-        if (VSFS_DEBUG()) {
-            if (fp) {
-                if (VSFS_DEBUG() > 2) {
-                    cerr << fullpath << " SUCCESS !!!" << endl;
-                }
-            } else {
-                cerr << output << " NOT FOUND !!!" << endl;
+        if (!found) {
+            try {
+                fullpath = boost::filesystem::canonical(filename_path);
+                found = true;
+            } catch (boost::filesystem::filesystem_error& e) {
+                found = false;
             }
+        }
+        if (!found) {
+            try {
+                fullpath = boost::filesystem::canonical(filename_path, datadir_path);
+                found = true;
+            } catch (boost::filesystem::filesystem_error& e) {
+                found = false;
+            }
+        }
+
+        if (found) {
+            std::string fullpath_string = fullpath.string();
+            const char* fullpath_c_string = fullpath_string.c_str();
+            fp = fopen(fullpath_c_string, mode);
+            if (!fp) {
+                VS_LOG(error, (boost::format("File '%1%' appears to exist, but we weren't able to open it.") % fullpath_string));
+            } else if (VSFS_DEBUG() > 2) {
+                VS_LOG(debug, (boost::format("SUCCESS opening file '%1%'") % fullpath_string));
+            }
+        } else {
+            VS_LOG(error, (boost::format("File '%1%' NOT FOUND") % filename));
         }
     } else {
-        fp = fopen(fullpath.c_str(), mode);
-        if (fp) {
-            if (VSFS_DEBUG() > 2) {
-                cerr << fullpath << " opened for writing SUCCESS !!!" << endl;
+        try {
+            fullpath = boost::filesystem::canonical(filename_path, homedir_path);
+            found = true;
+        } catch (boost::filesystem::filesystem_error& e) {
+            found = false;
+        }
+        if (found) {
+            std::string fullpath_string = fullpath.string();
+            const char* fullpath_c_string = fullpath_string.c_str();
+            fp = fopen(fullpath_c_string, mode);
+            if (fp) {
+                if (VSFS_DEBUG() > 2) {
+                    VS_LOG(debug, (boost::format("'%1%' opened for writing SUCCESS") % fullpath_string));
+                }
             }
-        } else if (VSFS_DEBUG()) {
-            cerr << fullpath << " FAILED !!!" << endl;
+        } else {
+            VS_LOG(error, (boost::format("Opening '%1%' FAILED !!!") % filename));
         }
     }
     return fp;
-
-    return nullptr;
 }
 
 size_t vs_read(void *ptr, size_t size, size_t nmemb, FILE *fp) {
@@ -708,10 +735,10 @@ void LoadConfig(string subdir) {
 
     //Now check if there is a data directory specified in it
     //NOTE : THIS IS NOT A GOOD IDEA TO HAVE A DATADIR SPECIFIED IN THE CONFIG FILE
-    if (!game_options()->datadir.empty()) {
+    if (!vs_options::instance().datadir.empty()) {
         //We config_file_found a path to data in config file
-        VS_LOG(important_info, (boost::format("DATADIR - Found a datadir in config, using : %1%") % game_options()->datadir));
-        datadir = game_options()->datadir;
+        VS_LOG(important_info, (boost::format("DATADIR - Found a datadir in config, using : %1%") % vs_options::instance().datadir));
+        datadir = vs_options::instance().datadir;
         data_dir_found = true;
     } else {
         if (true == legacy_data_dir_mode) {
@@ -741,14 +768,14 @@ void InitMods() {
     struct dirent **dirlist;
     //new config program should insert hqtextures variable
     //with value "hqtextures" in data section.
-    if (!game_options()->hqtextures.empty()) {
+    if (!vs_options::instance().hqtextures.empty()) {
         //HQ Texture dir sits alongside data dir.
         selectcurrentdir = datadir + "/..";
         boost::filesystem::path p(selectcurrentdir);
         for (boost::filesystem::directory_entry& entry : boost::filesystem::directory_iterator(p)) {
             const boost::filesystem::path& filename = entry.path().filename();
             const std::string filename_string = filename.string();
-            if (filename_string == game_options()->hqtextures && is_directory(entry.status())) {
+            if (filename_string == vs_options::instance().hqtextures && is_directory(entry.status())) {
                 curpath.clear();
                 curpath.append(selectcurrentdir);
                 curpath.append("/");
@@ -762,7 +789,7 @@ void InitMods() {
 //        if (ret >= 0) {
 //            while (ret--) {
 //                string dname(dirlist[ret]->d_name);
-//                if (dname == game_options()->hqtextures) {
+//                if (dname == vs_options::instance().hqtextures) {
 //                    curpath.clear();
 //                    curpath.append(selectcurrentdir);
 //                    curpath.append("/");
@@ -881,21 +908,21 @@ void InitPaths(std::string conf, std::string subdir) {
         SubDirectories.push_back(vec);
     }
 
-    game_options()->init();
+    vs_options::instance().init();
 
-    sharedsectors = game_options()->sectors;
-    sharedcockpits = game_options()->cockpits;
-    shareduniverse = game_options()->universe_path;
-    sharedanims = game_options()->animations;
-    sharedvideos = game_options()->movies;
-    sharedsprites = game_options()->sprites;
-    savedunitpath = game_options()->serialized_xml;
-    sharedtextures = game_options()->sharedtextures;
-    sharedsounds = game_options()->sharedsounds;
-    sharedmeshes = game_options()->sharedmeshes;
-    sharedunits = game_options()->sharedunits;
-    aidir = game_options()->ai_directory;
-    universe_name = game_options()->galaxy;
+    sharedsectors = vs_options::instance().sectors;
+    sharedcockpits = vs_options::instance().cockpits;
+    shareduniverse = vs_options::instance().universe_path;
+    sharedanims = vs_options::instance().animations;
+    sharedvideos = vs_options::instance().movies;
+    sharedsprites = vs_options::instance().sprites;
+    savedunitpath = vs_options::instance().serialized_xml;
+    sharedtextures = vs_options::instance().sharedtextures;
+    sharedsounds = vs_options::instance().sharedsounds;
+    sharedmeshes = vs_options::instance().sharedmeshes;
+    sharedunits = vs_options::instance().sharedunits;
+    aidir = vs_options::instance().ai_directory;
+    universe_name = vs_options::instance().galaxy;
 
     //Setup the directory lists we know about - note these are relative paths to datadir or homedir
     //----- THE Directories vector contains the resource/volume files name without extension or the main directory to files
@@ -944,9 +971,9 @@ void InitPaths(std::string conf, std::string subdir) {
     Directories[PythonFile] = "bases";
     Directories[AccountFile] = "accounts";
 
-    SIMULATION_ATOM = game_options()->simulation_atom;
+    SIMULATION_ATOM = vs_options::instance().simulation_atom;
     simulation_atom_var = SIMULATION_ATOM;
-    AUDIO_ATOM = game_options()->audio_atom;
+    AUDIO_ATOM = vs_options::instance().audio_atom;
     audio_atom_var = AUDIO_ATOM;
     VS_LOG(info, (boost::format("SIMULATION_ATOM: %1%") % SIMULATION_ATOM));
 
@@ -969,7 +996,7 @@ void InitPaths(std::string conf, std::string subdir) {
     //NOTE : UniverseFiles cannot use volumes since some are needed by python
     //Also : Have to try with systems, not sure it would work well
     //Setup the use of volumes for certain VSFileType
-    volume_format = game_options()->volume_format;
+    volume_format = vs_options::instance().volume_format;
     if (volume_format == "vsr") {
         q_volume_format = vfmtVSR;
     } else if (volume_format == "pk3") {
